% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atp.R
\name{atp_parse_page}
\alias{atp_parse_page}
\title{Parse a single ATP page}
\usage{
atp_parse_page(page)
}
\arguments{
\item{page}{Scraped page object (\code{xml_document}).}
}
\value{
A tibble.
}
\description{
Parse a single ATP page
}
\examples{
\dontrun{
html <- poddr:::poddr_get("https://atp.fm", as = "html", query = list(page = 1))
atp_parse_page(html)
}
}
