% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gather_people}
\alias{gather_people}
\title{Gather episode datasets by people}
\usage{
gather_people(episodes, people_cols = c("host", "guest"))
}
\arguments{
\item{episodes}{A tbl containing \code{host} and \code{guest} columns, with names
separated by \verb{;}.}

\item{people_cols}{FOr The Incomparable, use the default \code{c("host", "guest")},
for relay.fm, there's only a \code{"host"} column.}
}
\value{
A tibble with new columns \code{"role"} and \code{"person"}, one row per person.
}
\description{
Gather episode datasets by people
}
\examples{
\dontrun{
incomparable <- incomparable_get_episodes(incomparable_get_shows())
incomparable_wide <- gather_people(incomparable)
}
}
