% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incomparable.R
\name{incomparable_get_episodes}
\alias{incomparable_get_episodes}
\title{Retrieve all episodes for The Incomparable shows}
\usage{
incomparable_get_episodes(incomparable_shows, cache = TRUE)
}
\arguments{
\item{incomparable_shows}{Dataset of shows as returned by \code{incomparable_get_shows()}.}

\item{cache}{(\code{logical(1)}) Toggle the httr2 HTTP cache. Default \code{TRUE}.}
}
\value{
A tibble.
}
\description{
Retrieve all episodes for The Incomparable shows
}
\examples{
\dontrun{
shows <- incomparable_get_shows()
incomparable_get_episodes(shows)
}
}
