% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incomparable.R
\name{incomparable_get_shows}
\alias{incomparable_get_shows}
\title{Retrieve all The Incomparable shows}
\usage{
incomparable_get_shows(cache = TRUE)
}
\arguments{
\item{cache}{(\code{logical(1)}) Toggle the httr2 HTTP cache. Default \code{TRUE}.}
}
\value{
A tibble with columns \code{show}, \code{stats_url}, \code{archive_url}, \code{status}.
}
\description{
Retrieve all The Incomparable shows
}
\examples{
\dontrun{
incomparable_get_shows()
}
}
