% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incomparable.R
\name{incomparable_get_subcategories}
\alias{incomparable_get_subcategories}
\title{Extract subcategory index for given show}
\usage{
incomparable_get_subcategories(
  archive_url = "https://www.theincomparable.com/gameshow/archive/",
  cache = TRUE
)
}
\arguments{
\item{archive_url}{E.g. \code{"https://www.theincomparable.com/theincomparable/archive/"}.}

\item{cache}{(\code{logical(1)}) Toggle the httr2 HTTP cache. Default \code{TRUE}.}
}
\value{
A tibble with subcategory links and category names.
}
\description{
Extract subcategory index for given show
}
\examples{
\dontrun{
incomparable_get_subcategories("https://www.theincomparable.com/gameshow/archive/")
}
}
