% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incomparable.R
\name{incomparable_parse_archive}
\alias{incomparable_parse_archive}
\title{Parse a show's archive page on The Incomparable website}
\usage{
incomparable_parse_archive(archive_url, cache = TRUE)
}
\arguments{
\item{archive_url}{E.g. \code{"https://www.theincomparable.com/theincomparable/archive/"}.}

\item{cache}{(\code{logical(1)}) Toggle the httr2 HTTP cache. Default \code{TRUE}.}
}
\value{
A tibble.
}
\description{
Parse a show's archive page on The Incomparable website
}
\examples{
\dontrun{
incomparable_parse_archive("https://www.theincomparable.com/gameshow/archive/")
}
}
