% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incomparable.R
\name{incomparable_parse_stats}
\alias{incomparable_parse_stats}
\title{Parse The Incomparable stats.txt files}
\usage{
incomparable_parse_stats(stats_url, cache = TRUE)
}
\arguments{
\item{stats_url}{URL to the \code{stats.txt}.}

\item{cache}{(\code{logical(1)}) Toggle the httr2 HTTP cache. Default \code{TRUE}.}
}
\value{
A tibble.
}
\description{
Parse The Incomparable stats.txt files
}
\examples{
\dontrun{
incomparable_parse_stats("https://www.theincomparable.com/salvage/stats.txt")
}
}
