% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label_n}
\alias{label_n}
\title{Convenience function to display N}
\usage{
label_n(x, brackets = FALSE)
}
\arguments{
\item{x}{Data or singular value.}

\item{brackets}{Set \code{TRUE} to enclose result in \verb{( )}.}
}
\value{
A character of length 1.
}
\description{
Convenience function to display N
}
\examples{
label_n(100)
label_n(tibble(x = 1:10, y = 1:10), brackets = TRUE)
}
