% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poddr-package.R
\docType{package}
\name{poddr-package}
\alias{poddr}
\alias{poddr-package}
\title{poddr: Collect Metadata for Selected Podcasts}
\description{
Collecting all the data, but just for The Incomparable, Relay.fm and ATP.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jemus42/poddr}
  \item Report bugs at \url{https://github.com/jemus42/poddr/issues}
}

}
\author{
\strong{Maintainer}: Lukas Burk \email{lukas@tadaa-data.de} (\href{https://orcid.org/0000-0001-7528-3795}{ORCID})

}
\keyword{internal}
