% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relay.R
\name{relay_get_shows}
\alias{relay_get_shows}
\title{Retrieve all relay.fm shows}
\usage{
relay_get_shows()
}
\value{
A tibble with one row for each show
}
\description{
Parses the show overview page and returns a tibble of show names
with corresponding feed URLs, which in turn can then be passed to
\code{relay_parse_feed()} individually.
}
\examples{
\dontrun{
relay_get_shows()
}
}
