% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relay.R
\name{relay_parse_feed}
\alias{relay_parse_feed}
\title{Parse a relay.fm show feed}
\usage{
relay_parse_feed(url)
}
\arguments{
\item{url}{A show's feed URL, e.g. \code{"https://www.relay.fm/ungeniused/feed"}.}
}
\value{
A tibble.
}
\description{
Parse a relay.fm show feed
}
\examples{
\dontrun{
relay_parse_feed(url = "https://www.relay.fm/ungeniused/feed")
}
}
