% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atp.R
\name{atp_get_episodes}
\alias{atp_get_episodes}
\title{Get all ATP episodes}
\usage{
atp_get_episodes(page_limit = NULL)
}
\arguments{
\item{page_limit}{Number of pages to scrape, from newest to oldest episode.
Page 1 contains the 5 most recent episodes, and subsequent pages contain 50
episodes per page. As of December 2020, there are 10 pages total.
Pass \code{NULL} (default) to get all pages.}
}
\value{
A tibble.
}
\description{
Get all ATP episodes
}
\examples{
\dontrun{
atp <- atp_get_episodes(page_limit = 2)
}
}
