% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atp.R
\name{atp_parse_page}
\alias{atp_parse_page}
\title{Parse a single ATP page}
\usage{
atp_parse_page(page)
}
\arguments{
\item{page}{Scraped page.}
}
\value{
A tibble.
}
\description{
Parse a single ATP page
}
\examples{
\dontrun{
session <- polite::bow(url = "https://atp.fm")
page <- polite::scrape(session, query = list(page = 1))
atp_parse_page(page)
}
}
