% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-caching.R
\name{cache_podcast_data}
\alias{cache_podcast_data}
\title{Cache episode data}
\usage{
cache_podcast_data(x, dir = "data_cache", filename = NULL, csv = TRUE)
}
\arguments{
\item{x}{Object to cache.}

\item{dir}{\verb{["data_cache"]} Directory to save data to.}

\item{filename}{Optional filename sans extension, if not specified the name of \code{x} is used.}

\item{csv}{If \code{TRUE} (default), also saves a CSV file with the same base name.}
}
\value{
Nothing
}
\description{
Cache episode data
}
\examples{
\dontrun{
atp_new <- atp_get_episodes(page_limit = 1)
cache_podcast_data(atp_new, csv = FALSE)
}
}
