% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-caching.R
\name{cache_podcast_data}
\alias{cache_podcast_data}
\title{Cache episode data}
\usage{
cache_podcast_data(x, dir = "data", filename = NULL, csv = TRUE)
}
\arguments{
\item{x}{Object to cache}

\item{dir}{Directory to save data to}

\item{filename}{Optional filename, if not specified the name of \code{x}.}

\item{csv}{If \code{TRUE}, also saves a CSV file with the same base name.}
}
\value{
Nothing
}
\description{
Cache episode data
}
\examples{
\dontrun{
if (FALSE) cache_podcast_data()
}
}
