% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incomparable.R
\name{incomparable_get_episodes}
\alias{incomparable_get_episodes}
\title{Get all The Incomparable shows}
\usage{
incomparable_get_episodes(incomparable_shows)
}
\arguments{
\item{incomparable_shows}{Dataset of shows with title and URLs}
}
\value{
A tibble.
}
\description{
Get all The Incomparable shows
}
\examples{
\dontrun{
incomparable <- incomparable_get_episodes(incomparable_get_shows())
}
}
