% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incomparable.R
\name{incomparable_get_episodes}
\alias{incomparable_get_episodes}
\title{Retrieve all episodes for The Incomparable shows}
\usage{
incomparable_get_episodes(incomparable_shows)
}
\arguments{
\item{incomparable_shows}{Dataset of shows with title and URLs as returned by
\code{incomparable_get_shows()}.}
}
\value{
A tibble with one row per episode.
}
\description{
This combines \code{incomparable_parse_stats()} and \code{incomparable_parse_archive()}
to retrieve full episode information including host/guest, durations
including seconds, podcast subcategories and topics.
Use sparingly to limit unecessarily hammering the poor webserver!
}
\examples{
\dontrun{
incomparable <- incomparable_get_episodes(incomparable_get_shows())
}
}
