% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incomparable.R
\name{incomparable_get_shows}
\alias{incomparable_get_shows}
\title{Get The Incomparable shows}
\usage{
incomparable_get_shows()
}
\value{
A tibble.
}
\description{
Get The Incomparable shows
}
\examples{
\dontrun{
incomparable_get_shows()
}
}
