% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incomparable.R
\name{incomparable_get_shows}
\alias{incomparable_get_shows}
\title{Retrieve all The Incomparable shows}
\usage{
incomparable_get_shows(cache = TRUE)
}
\arguments{
\item{cache}{(\code{logical(1)}) Set to \code{FALSE} to disable caching.}
}
\value{
A tibble with following columns:

\if{html}{\out{<div class="sourceCode">}}\preformatted{Columns: 4
$ show        <chr>
$ stats_url   <glue>
$ archive_url <glue>
$ status      <chr>
}\if{html}{\out{</div>}}
}
\description{
Parses the show overview page and returns a tibble of show names
with corresponding URLs, which in turn can then be passed to
\code{incomparable_parse_archive()} and \code{incomparable_parse_stats()} individually.
}
\examples{
\dontrun{
incomparable_get_shows()
}
}
