% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incomparable.R
\name{incomparable_get_shows}
\alias{incomparable_get_shows}
\title{Get The Incomparable shows}
\usage{
incomparable_get_shows(url = "https://www.theincomparable.com/shows/")
}
\arguments{
\item{url}{Show index URL: \code{"https://www.theincomparable.com/shows/"}.}
}
\value{
A tibble.
}
\description{
Get The Incomparable shows
}
\examples{
\dontrun{
incomparable_get_shows()
}
}
