% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incomparable.R
\name{incomparable_parse_archive}
\alias{incomparable_parse_archive}
\title{Parse a show's archive page on The Incomparable website}
\usage{
incomparable_parse_archive(archive_url)
}
\arguments{
\item{archive_url}{E.g. \code{"https://www.theincomparable.com/theincomparable/archive/"}.}
}
\value{
A tibble.
}
\description{
Parse a show's archive page on The Incomparable website
}
\examples{
\dontrun{
archive_url <- "https://www.theincomparable.com/salvage/archive/"
incomparable_parse_archive(archive_url)
}
}
