% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incomparable.R
\name{incomparable_parse_archive}
\alias{incomparable_parse_archive}
\title{Parse a show's archive page on The Incomparable website}
\usage{
incomparable_parse_archive(archive_url)
}
\arguments{
\item{archive_url}{E.g.
\code{"https://www.theincomparable.com/theincomparable/archive/"}.}
}
\value{
A tibble.
}
\description{
Retrieves all episodes for one or more shows passed as a tibble.
The archive page \emph{does not} include full duration information, as it is
limited to hours and minutes. Use \code{incomparable_parse_stats()} for
accurate episode durations.
}
\examples{
\dontrun{
archive_url <- "https://www.theincomparable.com/salvage/archive/"
incomparable_parse_archive(archive_url)
}
}
