% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incomparable.R
\name{incomparable_parse_stats}
\alias{incomparable_parse_stats}
\title{WIP Parse The Incomparable stats.txt files}
\usage{
incomparable_parse_stats(stats_url)
}
\arguments{
\item{stats_url}{URL to the \code{stats.txt}.}
}
\value{
A tibble.
}
\description{
WIP Parse The Incomparable stats.txt files
}
\examples{
\dontrun{
stats_url <- "https://www.theincomparable.com/theincomparable/stats.txt"
incomparable_parse_stats(stats_url)
}
}
