% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incomparable.R
\name{incomparable_parse_stats}
\alias{incomparable_parse_stats}
\title{Parse The Incomparable stats.txt files}
\usage{
incomparable_parse_stats(stats_url)
}
\arguments{
\item{stats_url}{URL to the \code{stats.txt}, e.g.
\code{"https://www.theincomparable.com/salvage/stats.txt"}.}
}
\value{
A tibble.
}
\description{
The \code{stats.txt} files have a slightly different format, especially the
host/guest information may differ from what is returned by
\code{incomparable_parse_archive()}, which implicitly assumes the first person
mentioned to be the host of the episode. However, this data source
does not include podcast subcategories (e.g. "Old Movie Club") or
topic information, which is only available on the archive page.
}
\examples{
\dontrun{
stats_url <- "https://www.theincomparable.com/salvage/stats.txt"
incomparable_parse_stats(stats_url)
}
}
