% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_duration}
\alias{parse_duration}
\title{Converting HH:MM:SS or MM:SS to \code{hms}}
\usage{
parse_duration(x)
}
\arguments{
\item{x}{A duration}
}
\value{
A numeric of durations in \code{hms::hms()}.
}
\description{
Converting HH:MM:SS or MM:SS to \code{hms}
}
\examples{
parse_duration("32:12")
parse_duration("32:12:04")
}
