% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relay.R
\name{relay_get_episodes}
\alias{relay_get_episodes}
\title{Collect all relay.fm shows}
\usage{
relay_get_episodes(relay_shows)
}
\arguments{
\item{relay_shows}{A tibble of shows, from \code{relay_get_shows()}.}
}
\value{
A tibble.
}
\description{
Collect all relay.fm shows
}
\examples{
\dontrun{
relay <- relay_get_episodes(relay_get_shows())
}
}
