% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relay.R
\name{relay_get_shows}
\alias{relay_get_shows}
\title{Collect all relay.fm shows}
\usage{
relay_get_shows(url = "https://www.relay.fm/shows")
}
\arguments{
\item{url}{Show overview page: \code{"https://www.relay.fm/shows"}.}
}
\value{
A tibble with one row for each show
}
\description{
Collect all relay.fm shows
}
\examples{
\dontrun{
relay_get_shows()
}
}
