% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-caching.R
\name{update_cached_data}
\alias{update_cached_data}
\title{Update and cache data locally}
\usage{
update_cached_data(dir = "data_cache")
}
\arguments{
\item{dir}{\verb{["data_cache"]} Directory path to save cached data to.}
}
\value{
Nothing
}
\description{
Update and cache data locally
}
\examples{
\dontrun{
if (FALSE) {
update_cached_data()
}
}
}
